[Feat] VoteResultResponse에 rank 정보 추가 응답
