+++ app/src/main/java/fc/be/app/domain/vote/service/dto/response/VoteResultResponse.java
@@ -1,9 +1,12 @@
 package fc.be.app.domain.vote.service.dto.response;
 
+import fc.be.app.domain.member.entity.Member;
 import fc.be.app.domain.vote.entity.Candidate;
+import fc.be.app.domain.vote.entity.Vote;
 import fc.be.app.domain.vote.service.dto.response.vo.MemberProfile;
 import fc.be.app.domain.vote.service.dto.response.vo.PlaceInfo;
 
+import java.util.Comparator;
 import java.util.List;
 
 public record VoteResultResponse(
@@ -21,9 +24,10 @@ public record VoteResultResponse(
             List<MemberProfile> votedMemberProfiles,
             String tagline,
             boolean amIVote,
-            int voteCount
+            int voteCount,
+            int rank
     ) {
-        public static CandidateResultResponse of(Long memberId, Candidate candidate) {
+        public static CandidateResultResponse of(Long memberId, Candidate candidate, int rank) {
             return new CandidateResultResponse(
                     candidate.getId(),
                     PlaceInfo.of(candidate.getPlace()),
@@ -34,7 +38,32 @@ public record VoteResultResponse(
                     candidate.getTagline(),
                     candidate.getVotedMember().stream()
                             .anyMatch(votedMember -> votedMember.isMemberVote(memberId)),
-                    candidate.getVotedCount());
+                    candidate.getVotedCount(),
+                    rank);
         }
+    }
+
+    public static VoteResultResponse of(Vote vote, Member member) {
+        List<Integer> sortedCount = vote.getCandidates().stream()
+                .map(Candidate::getVotedCount)
+                .distinct()
+                .sorted(Comparator.reverseOrder())
+                .toList();
+
+        return new VoteResultResponse(
+                vote.getId(),
+                vote.getTitle(),
+                vote.getStatus().getDescription(),
+                MemberProfile.of(vote.getOwner()),
+                vote.getCandidates()
+                        .stream()
+                        .sorted(sortByVotedCount())
+                        .map(candidate -> CandidateResultResponse.of(member.getId(), candidate, sortedCount.indexOf(candidate.getVotedCount()) + 1))
+                        .toList()
+        );
+    }
+
+    private static Comparator<Candidate> sortByVotedCount() {
+        return Comparator.comparingInt(Candidate::getVotedCount).reversed();
     }
 }
